web_app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
databases[asyncpg]==0.9.0
asyncpg==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.12
Pillow==10.4.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
torchvision==0.18.1+cpu
requests==2.32.3
